A quite long script that generates a random e-mail signature file. The idea is to have a standard fixed signature file (e.g., your name and office), and a set of variable files, at least one for each day of month (but even for each day of year) with a random phrase organized into a directory like: ~/.signatureSelector/01.txt ~/.signatureSelector/02.txt ~/.signatureSelector/03.txt ... ~/.signatureSelector/31.txt
(or for the each-day-of-the-year)

~/.signatureSelector/01/01.txt
..
~/.signatureSelector/01/31.txt
~/.signatureSelector/02/01.txt
...
~/.signatureSelector/01/29.txt
...

The script will generate each time is invoked (e.g., at the user login)
a file with the fixed one merged with the variable one, so that each time
your mail client looks at the signature file it will find it different.

As a sidenote, I wrote this script in 2004 while on an airplane coming back
from Las Vegas (PPPJ conference).